requirements.txt
mido[ports-rtmidi]
openai
python-dotenv
midi2audio